+++ CHANGES.rst
@@ -4,7 +4,9 @@ Changelog
 2.0.0 (2022-09-19)
 ------------------
 
-- Changed exception from CannotBeChangeException to CannotBeChanged
+- Rename CannotBeChangeException to CannotBeChanged
+- Rename InvariantReturnValueException to InvariantMustReturnBool
+
 
 1.5.0 (2020-12-07)
 ------------------